test(flake)(#219): deterministic flush-settle waits across profile tests
Replace fixed-time `setTimeout(r, N)` waits after `save()` / `scheduleFlush*()`
with the deterministic `waitForFlushSettled(provider)` helper across the
profile test suite. Issue #219 surfaced a tombstone-write race in the G4
no-token-loss test (Run 1/10 of the #217 stress) caused by a 200ms budget
that under full-suite worker contention was not always enough for the
pin → bundle → operational-state writes to complete. Same anti-pattern
existed across 7 profile test files.

Promoted `waitForFlushSettled` from `pointer-monotonicity.test.ts` to a
shared helper at `tests/helpers/profile/waitForFlushSettled.ts`. The
helper polls the provider's private `flushPromise` / `flushTimer` fields
and returns only when both are null across a 5ms stability window — or
when the in-flight flush settles (including normal rejections like
`POINTER_MONOTONICITY_VIOLATION`).

Files touched:
- tests/helpers/profile/waitForFlushSettled.ts (new shared helper)
- tests/integration/profile/no-token-loss-tombstones.test.ts (target of #219)
- tests/unit/profile/pointer-monotonicity.test.ts (import shared helper)
- tests/unit/profile/wave-g7-prereq.test.ts
- tests/unit/profile/per-entry-key-writers.test.ts
- tests/unit/profile/profile-token-storage-flush-serialization.test.ts
- tests/unit/profile/profile-token-storage-no-data-flush.test.ts
- tests/unit/profile/profile-token-storage-provider.test.ts

In `flush-serialization.test.ts` the 20ms mid-flight waits are kept as
plain timers — they intentionally probe the in-flight state (e.g.
`expect(tracker.inFlight).toBe(1)`); replacing them would defeat the
test. Only the end-of-test "wait for all flushes to settle" waits were
converted. Similarly, the post-`_triggerReplication` event-handler wait
in `profile-token-storage-provider.test.ts:1178` is left alone — it
waits for a synchronous event push, not a flush body.

Verification: full-scope 10-run stress (461 files, 7915 tests each),
10/10 clean. Touched-files 5-run stress, 5/5 clean.

Refs: #219, #217 (stress-run surface), #215 (original pattern fix in
PR #216).
